hcp experiment updates for registration

+++ dataset/hcp/dwi/dwi.py
@@ -5,9 +5,10 @@ import shutil
 from util.logging import get_logger, set_logger
 from fwk.config import Config
 from util.path import absolute_path
+from dataset.hcp.dwi.mask import get_mask
 
 from dataset.hcp.hcp import HcpDiffusionDatabase
-from dataset.hcp.dwi.registration import find_prealign_affine, find_warp, apply_warp_on_volumes
+from dataset.hcp.dwi.registration import find_prealign_affine, find_warp, apply_warp_on_volumes, find_rigid_affine
 
 from dipy.io.image import load_nifti, save_nifti
 from dipy.reconst.csdeconv import ConstrainedSphericalDeconvModel
@@ -56,19 +57,36 @@ class HcpDwiProcessor:
         self.logger.info(f'processing subject {subject}')
 
         self.logger.info(f'registering subject {subject}')
-        self.register(subject)
+        # self.register_nonlinear(subject)
+        self.register_rigid(subject)
 
-        self.logger.info(f'fitting dti for subject {subject}')
-        self.fit_dti(subject)
+        # self.logger.info(f'fitting dti for subject {subject}')
+        # self.fit_dti(subject)
 
-        self.logger.info(f'fitting odf for subject {subject}')
-        self.fit_odf(subject)
+        # self.logger.info(f'fitting odf for subject {subject}')
+        # self.fit_odf(subject)
 
         if delete_folders is True:
             self.logger.info(f'cleaning up for subject {subject}')
             self.clean(subject)
 
-    def register(self, subject):
+    def register_rigid(self, subject):
+
+        static_url = self._template_fa_url()
+        static, static_affine = load_nifti(static_url)
+
+        moving_url = self._get_moving_fa(subject)
+
+        moving, moving_affine = load_nifti(moving_url)
+
+        rigid_affine = find_rigid_affine(static, static_affine,  moving, moving_affine)
+
+        rigid_affine_url = self._url_moving_dwi(subject, 'rigid_affine.nii.gz')
+        save_nifti(rigid_affine_url, rigid_affine.affine, rigid_affine.affine)
+
+        # find linear registration transformation
+
+    def register_nonlinear(self, subject):
 
         static_url = self._template_fa_url()
         static, static_affine = load_nifti(static_url)
@@ -96,7 +114,11 @@ class HcpDwiProcessor:
 
         self._save_registered_files(subject, warped_data, warped_fa, template_mask, template_affine)
 
-    def _save_registered_files(self, subject, warped_data, warped_fa, warped_mask, warped_affine):
+    def _save_registered_files(self, subject, warped_data, warped_fa, warped_mask, warped_affine, rigid_affine=None):
+
+        if rigid_affine is not None:
+            rigid_affine_url = self._url_registered_dwi(subject, 'rigid_affine.nii.gz')
+            save_nifti(rigid_affine_url, rigid_affine.affine, rigid_affine.affine)
 
         warped_fa_url = self._url_registered_dwi(subject, 'warped_fa.nii.gz')
         save_nifti(warped_fa_url, warped_fa, warped_affine)
@@ -129,13 +151,29 @@ class HcpDwiProcessor:
             'output': self._url_moving_dwi(subject, 'dti')
         }
 
-        self._perform_dti_fit(dti_params, save_tensor=False)
+        self._perform_dti_fit(dti_params, save_tensor=True)
+
+        moving_tensor_url = self._url_moving_dwi(subject, 'dti_tensor.*')
+        fslconvert_command_str = f'fslchfiletype NIFTI_GZ {moving_tensor_url}'
+        subprocess.run(fslconvert_command_str, shell=True, check=True)
 
         moving_fa_url = self._url_moving_dwi(subject, 'dti_FA.*')
         fslconvert_command_str = f'fslchfiletype NIFTI_GZ {moving_fa_url}'
         subprocess.run(fslconvert_command_str, shell=True, check=True)
-
         converted_moving_fa_url = self._url_moving_dwi(subject, 'dti_FA.nii.gz')
+
+        moving_md_url = self._url_moving_dwi(subject, 'dti_MD.*')
+        fslconvert_command_str = f'fslchfiletype NIFTI_GZ {moving_md_url}'
+        subprocess.run(fslconvert_command_str, shell=True, check=True)
+        converted_moving_md_url = self._url_moving_dwi(subject, 'dti_MD.nii.gz')
+
+        fa, affine_fa = load_nifti(converted_moving_fa_url)
+        md, _ = load_nifti(converted_moving_md_url)
+
+        wm_mask = get_mask(fa, md)
+        wm_mask_url = self._url_moving_dwi(subject, 'dti_WM.nii.gz')
+        affine_wm = affine_fa
+        save_nifti(wm_mask_url, wm_mask, affine_wm)
 
         return converted_moving_fa_url
 
@@ -202,7 +240,8 @@ class HcpDwiProcessor:
 
     def clean(self, subject):
         data_url = self._url_registered_dwi(subject, 'data.nii.gz')
-        os.remove(data_url)
+        if os.path.isfile(data_url):
+            os.remove(data_url)
 
     def _template_fa_url(self):
         return os.path.join(self.template_folder, self.template_file)
